+++ src/sendrecv.cu
@@ -35,7 +35,7 @@ testResult_t SendRecvInitData(struct threadArgs* args, ncclDataType_t type, nccl
   return testSuccess;
 }
 
-void SendRecvGetBw(size_t count, int typesize, double sec, double* algBw, double* busBw, int nranks) {
+void SendRecvGetBw(size_t count, size_t typesize, double sec, double* algBw, double* busBw, int nranks) {
   double baseBw = (double)(count * typesize) / 1.0E9 / sec;
 
   *algBw = baseBw;
